+++ util/util_dirs.py
@@ -1,0 +1,8 @@
+import os
+
+USER = 'sarperyurtseven'
+
+PLOTS_SAVE_PATH = f'/data/scratch/{USER}/results/plots'
+INJECTIONS = '/data/scratch/bariskurtkaya/dataset/NIRCAM/1386/injections'
+
+
